rename save file modals for clarity, because both will probably exist in parallel for a long time, as long as reports are generated on frontend which we have no intention to change
