+++ apps/admin/frontend/src/components/save_frontend_file_modal.tsx
@@ -46,7 +46,7 @@ enum ModalState {
   INIT = 'init',
 }
 
-export function SaveFileToUsb({
+export function SaveFrontendFileModal({
   onClose,
   generateFileContent,
   defaultFilename,
